requirements.txt
Django==2.2.10
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.12.4
Markdown==3.3.6
djangorestframework-camel-case==1.2.0
